implementação do gerador de traces

+++ System/Event.h
@@ -20,14 +20,14 @@
  *
  */
 struct Event {
-    int id{};                       // identificador do processo.
-    bool concluded{};               // indica se o processo foi concluído.
-    unsigned long tempoEspera{};    // tempo de espera do processo.
-    unsigned long tempoExecucao{};  // tempo de execução do processo.
-    unsigned long instanteCPU{};    // instante que foi enviado para execução na CPU.
-    unsigned long instanteDisco{};  // instante que foi enviado para execução no disco.
-    unsigned long instanteRede{};   // instante que foi enviado para execução na rede.
-    unsigned long instanteFinal{};  // instante de término do processo.
+    int id{};                  // identificador do processo.
+    bool concluded{};          // indica se o processo foi concluído.
+    unsigned tempoEspera{};    // tempo de espera do processo.
+    unsigned tempoExecucao{};  // tempo de execução do processo.
+    unsigned instanteCPU{};    // instante que foi enviado para execução na CPU.
+    unsigned instanteDisco{};  // instante que foi enviado para execução no disco.
+    unsigned instanteRede{};   // instante que foi enviado para execução na rede.
+    unsigned instanteFinal{};  // instante de término do processo.
 };
 
 #endif  // EVENT_H
